+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added: `defaultSeverity` configuration option.
 - Fixed: `time-no-imperceptible` correctly handles negative time.
+- Fixed: `function-calc-no-unspaced-operator` correctly interprets negative fractional numbers without leading zeros and those wrapped in parentheses.
 
 # 6.8.0